Move skeletons to the components folder
