+++ src/components/carbon/shared.js
@@ -1,63 +1,4 @@
-import styled, { keyframes, css } from 'styled-components';
-
-export const skeletonAnimation = keyframes`
-  0% {
-    width: 0%;
-    left: 0;
-    right: auto;
-    opacity: 0.3;
-  }
-
-  20% {
-    width: 100%;
-    left: 0;
-    right: auto;
-    opacity: 1;
-  }
-  
-  28% {
-    width: 100%;
-    left: auto;
-    right: 0;
-  }
-
-  51% {
-    width: 0%;
-    left: auto;
-    right: 0;
-  }
-
-  58% {
-    width: 0%;
-    left: auto;
-    right: 0;
-  }
-
-  82% {
-    width: 100%;
-    left: auto;
-    right: 0;
-  }
-
-  83% {
-    width: 100%;
-    left: 0;
-    right: auto;
-  }
-
-  96% {
-    width: 0%;
-    left: 0;
-    right: auto;
-  }
-
-  100% {
-    width: 0%;
-    left: 0;
-    right: auto;
-    opacity: 0.3;
-  }
-`;
+import styled, { css } from 'styled-components';
 
 export const bxLabel = css`
   font-size: 1.5rem;
